Move SDK resolution to separate file
